MS ML Pack/ml/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
xgboost==2.0.3
python-igraph==0.11.4
leidenalg==0.10.1
celery==5.3.6
